+++ spec/ruby/core/refinement/fixtures/classes.rb
@@ -1,0 +1,12 @@
+module RefinementSpec
+
+  module ModuleWithAncestors
+    include Module.new do
+      def indent(level)
+        " " * level + self
+      end
+    end
+  end
+end
+
+